File Structure change on phone auth service
