+++ api/server.js
@@ -46,13 +46,13 @@ if (config.util.getEnv("NODE_ENV") !== "test") {
 
 //Requiring Routes
 const readingRoutes = require("./routes/readingRoutes"),
-  phoneAuth = require("./auth/phoneAuth")
+  authServices = require("./auth/auth")
 
 //Using Routes
 app.use("/api/reading", readingRoutes);
 
 //Using phone auth routes
-app.use("/auth", phoneAuth);
+app.use("/auth", authServices);
 
 //Starting the server
 app.listen(port, err => {
